Implement initial "working" version
The main components are in place. bin/review script is runnable and
displays output in specified format. Need to verify correctness of
interesting item matches against spec.

This takes the approach of parsing each pull request's diff into an
object and then inspecting the diff's files and changed lines to see if
they're interesting.

ADDED
- octokit dependency
- bin/review script
- Review::CommandLine class
- Review::Review class
- Review::Diff, Review::File, Review::Hunk classes

TODO
- Make runnable offline
- Add tests
- Add mock data that exercises the spec
- Collect more metadata from diffs
- Collect interesting items and their locations; show in report
- Make files and patterns of interest configurable (via command line
  options and/or config file)
- Make interestingness configurable in general; allow strategy for
  finding interesting items to be pluggable